Migrate manage module to new structure
